Add new `NativeJob`
