+++ boa_engine/src/job.rs
@@ -3,6 +3,68 @@
 use crate::{prelude::JsObject, Context, JsResult, JsValue};
 use boa_gc::{Finalize, Trace};
 
+/// An ECMAScript [Job] closure.
+///
+/// The specification allows scheduling any [`NativeJob`] closure by the host into the job queue.
+/// However, custom jobs must abide to a list of requirements.
+///
+/// ### Requirements
+///
+/// - At some future point in time, when there is no running execution context and the execution
+/// context stack is empty, the implementation must:
+///     - Perform any host-defined preparation steps.
+///     - Invoke the Job Abstract Closure.
+///     - Perform any host-defined cleanup steps, after which the execution context stack must be empty.
+/// - Only one Job may be actively undergoing evaluation at any point in time.
+/// - Once evaluation of a Job starts, it must run to completion before evaluation of any other Job starts.
+/// - The Abstract Closure must return a normal completion, implementing its own handling of errors.
+///
+/// `NativeJob`s API differs slightly on the last requirement, since it allows closures returning
+/// [`JsResult`], but it's okay because `NativeJob`s are handled by the host anyways; a host could
+/// pass a closure returning `Err` and handle the error on [`JobQueue::run_jobs`], making the closure
+/// effectively run as if it never returned `Err`.
+///
+/// ## [`Trace`]?
+///
+/// `NativeJob` doesn't implement `Trace` because it doesn't need to; all jobs can only be run once
+/// and putting a [`JobQueue`] on a garbage collected object should definitely be discouraged.
+///
+/// On the other hand, it seems like this type breaks all the safety checks of the
+/// [`NativeFunction`] API, since you can capture any `Trace` variable into the closure... but it
+/// doesn't!
+/// The garbage collector doesn't need to trace the captured variables because the closures
+/// are always stored on the [`JobQueue`], which is always rooted, which means the captured variables
+/// are also rooted, allowing us to capture any variable in the closure for free!
+///
+///
+/// [Job]: https://tc39.es/ecma262/#sec-jobs
+/// [`NativeFunction`]: crate::native_function::NativeFunction
+pub struct NativeJob {
+    #[allow(clippy::type_complexity)]
+    f: Box<dyn FnOnce(&mut Context<'_>) -> JsResult<JsValue>>,
+}
+
+impl std::fmt::Debug for NativeJob {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("NativeJob").field("f", &"Closure").finish()
+    }
+}
+
+impl NativeJob {
+    /// Creates a new `NativeJob` from a closure.
+    pub fn new<F>(f: F) -> Self
+    where
+        F: FnOnce(&mut Context<'_>) -> JsResult<JsValue> + 'static,
+    {
+        Self { f: Box::new(f) }
+    }
+
+    /// Calls the native job with the specified [`Context`].
+    pub fn call(self, context: &mut Context<'_>) -> JsResult<JsValue> {
+        (self.f)(context)
+    }
+}
+
 /// `JobCallback` records
 ///
 /// More information:
